+++ lib/eve_online/esi/server_status.rb
@@ -1,34 +1,18 @@
+require 'forwardable'
+
 module EveOnline
   module ESI
     class ServerStatus < Base
+      extend Forwardable
+
       API_ENDPOINT = 'https://esi.tech.ccp.is/v1/status/?datasource=tranquility'.freeze
 
-      def as_json
-        {
-          start_time: start_time,
-          players: players,
-          server_version: server_version,
-          vip: vip
-        }
-      end
+      def_delegators :model, :as_json, :start_time, :players, :server_version, :vip
 
-      def start_time
-        start_time = response['start_time']
-
-        parse_datetime_with_timezone(start_time) if start_time
+      def model
+        Models::ServerStatus.new(response)
       end
-
-      def players
-        response['players']
-      end
-
-      def server_version
-        response['server_version']
-      end
-
-      def vip
-        response['vip']
-      end
+      memoize :model
 
       def scope; end